refactoring, new combatlog parser

+++ AmeisenBotX.Core/Fsm/Routines/SellItemsRoutine.cs
@@ -15,11 +15,11 @@ namespace AmeisenBotX.Core.Fsm.Routines
                 IWowInventoryItem itemToSell = item;
 
                 if (config.ItemSellBlacklist.Any(e => e.Equals(item.Name, StringComparison.OrdinalIgnoreCase))
-                    || (!config.SellGrayItems && item.ItemQuality == WowItemQuality.Poor)
-                    || (!config.SellWhiteItems && item.ItemQuality == WowItemQuality.Common)
-                    || (!config.SellGreenItems && item.ItemQuality == WowItemQuality.Uncommon)
-                    || (!config.SellBlueItems && item.ItemQuality == WowItemQuality.Rare)
-                    || (!config.SellPurpleItems && item.ItemQuality == WowItemQuality.Epic))
+                    || (!config.SellGrayItems && item.ItemQuality == (int)WowItemQuality.Poor)
+                    || (!config.SellWhiteItems && item.ItemQuality == (int)WowItemQuality.Common)
+                    || (!config.SellGreenItems && item.ItemQuality == (int)WowItemQuality.Uncommon)
+                    || (!config.SellBlueItems && item.ItemQuality == (int)WowItemQuality.Rare)
+                    || (!config.SellPurpleItems && item.ItemQuality == (int)WowItemQuality.Epic))
                 {
                     continue;
                 }
